+++ docs/src/conf.py
@@ -67,7 +67,10 @@ html_theme = "piccolo_theme"
 html_short_title = "Piccolo"
 html_show_sphinx = False
 globaltoc_maxdepth = 3
-html_theme_options = {"source_url": "https://github.com/piccolo-orm/piccolo/"}
+html_theme_options = {
+    "source_url": "https://github.com/piccolo-orm/piccolo/",
+    "banner_text": 'Piccolo v1 is now available! <a href="/en/latest/piccolo/v1/index.html">Learn more here</a>.',  # noqa: E501
+}
 
 # -- Options for HTMLHelp output ---------------------------------------------